Rename as UpsertCustomerIoProfile
